Move tap in UIAlertController to PostActionSheetTests
